+++ application/packages/web/src/client/features/inbox/components/inbox-article-card.tsx
@@ -1,0 +1,68 @@
+import { isArticleMedia } from '@trend-diary/domain/article/media'
+import { Button } from '@/client/components/shadcn/button'
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/client/components/shadcn/tooltip'
+import { MediaIcon, type MediaIconType } from '@/client/features/article'
+import type { Article } from '../hooks/use-unread-digestion'
+
+interface Props {
+  article: Article
+  onSkip: () => Promise<void>
+  onRead: () => Promise<void>
+  onLater: () => void
+}
+
+const toMediaType = (media: string): MediaIconType => {
+  if (isArticleMedia(media)) return media
+  return 'zenn'
+}
+
+export default function InboxArticleCard({ article, onSkip, onRead, onLater }: Props) {
+  return (
+    <div className='mt-2 flex flex-col rounded-xl border border-gray-200 bg-white p-5'>
+      <h2 className='flex h-[2lh] items-start gap-2 text-lg font-semibold text-gray-900 leading-relaxed'>
+        <span className='mt-0.5 shrink-0'>
+          <MediaIcon media={toMediaType(article.media)} size='md' />
+        </span>
+        <span className='line-clamp-2'>{article.title}</span>
+      </h2>
+      <div className='mt-2 flex h-[1lh] items-center gap-3 overflow-hidden text-sm text-gray-600'>
+        <span className='min-w-0 truncate'>著者: {article.author}</span>
+      </div>
+      <p className='mt-3 h-24 overflow-y-auto overflow-x-hidden whitespace-pre-wrap break-words pr-1 text-sm leading-relaxed text-gray-700 md:h-56'>
+        {article.description}
+      </p>
+      <div className='sticky bottom-0 -mx-5 mt-3 flex flex-wrap gap-2 border-t border-gray-200 bg-white/95 px-5 pt-3 pb-1 md:static md:mx-0 md:mt-4 md:border-0 md:bg-transparent md:p-0'>
+        <Tooltip>
+          <TooltipTrigger asChild={true}>
+            <Button type='button' variant='outline' onClick={onSkip}>
+              スキップ
+            </Button>
+          </TooltipTrigger>
+          <TooltipContent side='top' align='start' className='bg-primary/85'>
+            対象外にする
+          </TooltipContent>
+        </Tooltip>
+        <Tooltip>
+          <TooltipTrigger asChild={true}>
+            <Button type='button' onClick={onRead}>
+              読む
+            </Button>
+          </TooltipTrigger>
+          <TooltipContent side='top' align='start' className='bg-primary/85'>
+            読んで消化する
+          </TooltipContent>
+        </Tooltip>
+        <Tooltip>
+          <TooltipTrigger asChild={true}>
+            <Button type='button' variant='secondary' onClick={onLater}>
+              後で
+            </Button>
+          </TooltipTrigger>
+          <TooltipContent side='top' align='start' className='bg-primary/85'>
+            このセッション内であとで見る
+          </TooltipContent>
+        </Tooltip>
+      </div>
+    </div>
+  )
+}
