+++ application/packages/web/src/client/features/inbox/index.ts
@@ -1,1 +1,3 @@
+export { default as InboxArticleCard } from './components/inbox-article-card'
+export { default as InboxCompletionCard } from './components/inbox-completion-card'
 export { default as useUnreadDigestion, type Article } from './hooks/use-unread-digestion'
